requirements.txt
cython
h5py
keras
matplotlib
numpy>=1.14
cupy-cuda101
pillow
tqdm
tensorflow-gpu
scikit-learn
scikit-image
pandas
albumentations
typeguard